+++ src/semantic/checks/top_level_declaration.rs
@@ -23,7 +23,8 @@ impl SemanticCheck for Statement<'_> {
             Statement::Binding(b) => b.check_semantics(scope),
             Statement::FnDecl(f) => f.check_semantics(scope),
             Statement::Conditional(_) => unimplemented!("check conditional"),
-            Statement::ForLoop(_) => unimplemented!("check for loop"),
+            Statement::ForLoop(_) => unimplemented!("check for for loop"),
+            Statement::WhileLoop(_) => unimplemented!("check for while loop"),
         }
     }
 }