+++ src/syntax/ast/loops.rs
@@ -1,6 +1,6 @@
 use crate::lexic::token::Token;
 
-use super::Block;
+use super::{Block, Expression};
 
 #[derive(Debug)]
 pub struct ForLoop<'a> {
@@ -13,3 +13,15 @@ pub struct ForLoop<'a> {
     pub value: &'a Token,
     pub body: Block<'a>,
 }
+
+#[derive(Debug)]
+pub struct WhileLoop<'a> {
+    /// the start position of the
+    /// `for` keyword
+    pub loop_start: usize,
+    /// the position of the closing bracket
+    pub loop_end: usize,
+    pub condition: Expression<'a>,
+    pub body: Block<'a>,
+}
+