+++ src/syntax/parsers/mod.rs
@@ -6,3 +6,4 @@ pub mod for_loop;
 pub mod function_declaration;
 pub mod module;
 pub mod statement;
+pub mod while_loop;